requirements.txt
torch
torchvision
timm
transformers
datasets
sentencepiece
tokenizers
tqdm
pandas
numpy
scikit-learn
lightgbm
Pillow
